Update layer width with browser resize
Layers now take up the width of the canvas. Extra margins can be added
in later if users want more annotating space around models.

All the layers in the stack are redrawn whenever the browser window is
resized. To do this, the layers needed to store the variables they were
first drawn with. This will also apply later when they have other
properties.