Squash bug fixes after previous changes
This is the 1st commit message:

Fix errors caused by selection and resize

The reselection after the stack had been drawn was throwing errors where
it was refering to properties but no layer had been selected.

This is the commit message #2:

Fix draw order

Previous change prevented selected layer from being drawn at the top. So
the highlighting was being overlapped by layers above.

This is the commit message #3:

Fix bug with deselection

If the user clicked on the canvas or another non-layer element while a
layer was selected, then deselection would only partially complete.
Clicking on another element was setting the selected layer to undefined,
so most of the deselect method was being skipped after this if the user
pressed escape.